+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+- `v0.0.10` MultiSelectInput component
+  - Add `clearable` property to allow clearing the selected options
+  - Added tests for the new clearable functionality
+
 ## 1.11.12 - 2025-10-30
 ### Fixed
 - `v0.0.12` EntityInfoCardSection